🐛 fix(VSecM Helm Charts): remove ability to not use SCM (#1064)
This PR removes the useClusterSpiffeIDs and useSpireControllerManager
options from the helm charts.

SPIRE helm charts use SPIRE Controller Manager, and disabling it
is nontrivial. Also, ClusterSPIFFEIDs are the best way to manage SPIFFEIDs
in a Kubernetes cluster. — If we find a use case where these options are
necessary, or if there is a need from the community, we can modify the
code to let SPIRE install without SPIRE Controller Manager and bring those
flags back.

Signed-off-by: Volkan Özçelik <volkan.ozcelik@broadcom.com>