🌟 enhancement(VSecM): pointed example images to upstream (#1093)
update example image overrides to refer from the "latest"
tag on upstream instead of "localhost:5000". This will allow
people to follow the use cases and other examples in the documentation
easier as they won't have to build the project from source and
also they won't need to update the version numbers.

Since the version numbers in the repository often points to a future
version that has not been published yet, trying to install them
without updating the image-overrides was causing "image not found"
errors. This PR addresses that by always pointing the example
images to the `latest` upsream version.

Since these examples often do not change and are mature enough,
it does not quite matter where you pull them from as long as
you can download and install them.

This PR does not change other product images and image overrides,
it only affect the image overrides under the `./examples` folder.

Signed-off-by: Volkan Özçelik <volkan.ozcelik@broadcom.com>